Trait iced_core::clipboard::Clipboard

pub trait Clipboard {
    // Required methods
    fn read(&self) -> Option<String>;
    fn write(&mut self, contents: String);

    // Provided methods
    fn read_primary(&self) -> Option<String> { ... }
    fn write_primary(&mut self, _contents: String) { ... }
    fn read_data(&self, _mimes: Vec<String>) -> Option<(Vec<u8>, String)> { ... }
    fn write_data(
        &mut self,
        _contents: ClipboardStoreData<Box<dyn Send + Sync + AsMimeTypes + 'static>>
    ) { ... }
    fn read_primary_data(
        &self,
        _mimes: Vec<String>
    ) -> Option<(Vec<u8>, String)> { ... }
    fn write_primary_data(
        &mut self,
        _contents: ClipboardStoreData<Box<dyn Send + Sync + AsMimeTypes + 'static>>
    ) { ... }
}

A buffer for short-term storage and transfer within and between applications.

Required Methods

fn read(&self) -> Option<String>

Reads the current content of the Clipboard as text.

fn write(&mut self, contents: String)

Writes the given text contents to the Clipboard.

Provided Methods

fn read_primary(&self) -> Option<String>

Read the current content of the primary Clipboard as text.

fn write_primary(&mut self, _contents: String)

Writes the given text contents to the primary Clipboard.

fn read_data(&self, _mimes: Vec<String>) -> Option<(Vec<u8>, String)>

Consider using read_data instead Reads the current content of the Clipboard as text.

fn write_data( &mut self, _contents: ClipboardStoreData<Box<dyn Send + Sync + AsMimeTypes + 'static>> )

Writes the given contents to the Clipboard.

fn read_primary_data(&self, _mimes: Vec<String>) -> Option<(Vec<u8>, String)>

Consider using read_primary_data instead Reads the current content of the primary Clipboard as text.

fn write_primary_data( &mut self, _contents: ClipboardStoreData<Box<dyn Send + Sync + AsMimeTypes + 'static>> )

Writes the given text contents to the primary Clipboard.