iced_core

Function never

pub fn never<T>(never: Infallible) -> T

A function that can never be called.

This is useful to turn generic types into anything you want by coercing them into a type with no possible values.