Function iced_core::layout::flex::resolve

pub fn resolve<Message, Renderer>(
    axis: Axis,
    renderer: &Renderer,
    limits: &Limits,
    padding: Padding,
    spacing: f32,
    align_items: Alignment,
    items: &[Element<'_, Message, Renderer>],
    trees: &mut [Tree]
) -> Node
where
    Renderer: Renderer,

Computes the flex layout with the given axis and limits, applying spacing, padding and alignment to the items as needed.

It returns a new layout Node.