iced_core::layout

Function padded

pub fn padded(
    limits: &Limits,
    width: impl Into<Length>,
    height: impl Into<Length>,
    padding: impl Into<Padding>,
    layout: impl FnOnce(&Limits) -> Node,
) -> Node

Computes the Node that fits the Limits given some width, height, and Padding requirements and a closure that produces the content Node inside the given Limits.