iced_core::shell

Struct Shell

pub struct Shell<'a, Message> { /* private fields */ }

A connection to the state of a shell.

A Widget can leverage a Shell to trigger changes in an application, like publishing messages or invalidating the current layout.

Implementations

impl<'a, Message> Shell<'a, Message>

pub fn new( window: &'a dyn Window, waker: Waker, bus: &'a mut Bus<Message>, ) -> Self

Creates a new Shell with the provided buffer of messages.

pub fn local<'b, A>(&self, bus: &'b mut Bus<A>) -> Shell<'b, A>
where 'a: 'b,

Creates a new Shell from the current one with the given list of local messages.

pub fn window(&self) -> &'a dyn Window

Returns the Window of the Shell.

pub fn waker(&self) -> &Waker

Returns the Waker of the Shell.

pub fn is_empty(&self) -> bool

Returns true if the Shell contains no published messages

pub fn publish(&mut self, message: Message)

Publish the given Message for an application to process it.

pub fn publish_and_track(&mut self, message: Message) -> Tracking

Publish the given Message for an application to process it.

The returned Tracking can be used to determine if the message was processed.

pub fn capture_event(&mut self)

Marks the current event as captured. Prevents “event bubbling”.

A widget should capture an event when no ancestor should handle it.

pub fn event_status(&self) -> Status

Returns the current event::Status of the Shell.

pub fn is_event_captured(&self) -> bool

Returns whether the current event has been captured.

pub fn request_redraw(&mut self)

Requests a new frame to be drawn as soon as possible.

pub fn request_redraw_at(&mut self, redraw_request: impl Into<RedrawRequest>)

Requests a new frame to be drawn at the given window::RedrawRequest.

pub fn redraw_request(&self) -> RedrawRequest

Returns the request a redraw should happen, if any.

pub fn replace_redraw_request(shell: &mut Self, redraw_request: RedrawRequest)

Replaces the redraw request of the Shell; without conflict resolution.

This is useful if you want to overwrite the redraw request to a previous value. Since it’s a fairly advanced use case and should rarely be used, it is a static method.

pub fn read_clipboard(&mut self, kind: Kind)

Requests the runtime to read the clipboard contents expecting the given clipboard::Kind.

The runtime will produce a clipboard::Event::Read when the contents have been read.

pub fn write_clipboard(&mut self, content: impl Into<Content>)

Requests the runtime to write the given clipboard::Content to the clipboard.

The runtime will produce a clipboard::Event::Written when the contents have been written.

pub fn clipboard_mut(&mut self) -> &mut Clipboard

Returns the Clipboard requests of the Shell, mutably.

pub fn request_input_method<T: AsRef<str>>(&mut self, ime: &InputMethod<T>)

Requests the current InputMethod strategy.

Important: This request will only be honored by the Shell only during a window::Event::RedrawRequested.

pub fn input_method(&self) -> &InputMethod

Returns the current InputMethod strategy.

pub fn input_method_mut(&mut self) -> &mut InputMethod

Returns the current InputMethod strategy.

pub fn is_layout_invalid(&self) -> Option<Diff>

Returns whether the current layout is invalid or not.

pub fn invalidate_layout(&mut self)

Invalidates the current application layout.

The shell will relayout the application widgets.

pub fn invalidate_layout_with(&mut self, diff: Diff)

Invalidates the current application layout with the following Diff strategy.

pub fn revalidate_layout(&mut self, f: impl FnOnce(Diff))

Triggers the given function if the layout is invalid, cleaning it in the process.

pub fn are_widgets_invalid(&self) -> bool

Returns whether the widgets of the current application have been invalidated.

pub fn invalidate_widgets(&mut self)

Invalidates the current application widgets.

The shell will rebuild and relayout the widget tree.

pub fn merge<B>(&mut self, other: Shell<'_, B>, f: impl Fn(B) -> Message)

Merges the current Shell with another one by applying the given function to the messages of the latter.

This method is useful for composition.