Struct iced_core::Color

pub struct Color {
    pub r: f32,
    pub g: f32,
    pub b: f32,
    pub a: f32,
}

A color in the sRGB color space.

Fields
r: f32

Red component, 0.0 - 1.0

g: f32

Green component, 0.0 - 1.0

b: f32

Blue component, 0.0 - 1.0

a: f32

Transparency, 0.0 - 1.0

Implementations

impl Color

pub const BLACK: Color = _

The black color.

pub const WHITE: Color = _

The white color.

pub const TRANSPARENT: Color = _

A color with no opacity.

pub fn new(r: f32, g: f32, b: f32, a: f32) -> Color

Creates a new Color.

In debug mode, it will panic if the values are not in the correct range: 0.0 - 1.0

pub const fn from_rgb(r: f32, g: f32, b: f32) -> Color

Creates a Color from its RGB components.

pub const fn from_rgba(r: f32, g: f32, b: f32, a: f32) -> Color

Creates a Color from its RGBA components.

pub fn from_rgb8(r: u8, g: u8, b: u8) -> Color

Creates a Color from its RGB8 components.

pub fn from_rgba8(r: u8, g: u8, b: u8, a: f32) -> Color

Creates a Color from its RGB8 components and an alpha value.

pub fn from_linear_rgba(r: f32, g: f32, b: f32, a: f32) -> Self

Creates a Color from its linear RGBA components.

pub fn into_rgba8(self) -> [u8; 4]

Converts the Color into its RGBA8 equivalent.

pub fn into_linear(self) -> [f32; 4]

Converts the Color into its linear values.

pub fn invert(&mut self)

Inverts the Color in-place.

pub fn inverse(self) -> Color

Returns the inverted Color.

pub fn scale_alpha(self, factor: f32) -> Color

Scales the alpha channel of the Color by the given factor.