iced_core

Module text

Draw and interact with text.

Re-exports

pub use editor::Editor;

pub use highlighter::Highlighter;

pub use paragraph::Paragraph;

Modules

editor

Edit text.

highlighter

Highlight text.

paragraph

Draw paragraphs.

Structs

Highlight

A text highlight.

Span

A span of text.

Text

A paragraph.

Enums

Difference

The difference detected in some text.

Hit

The result of hit testing on text.

LineHeight

The height of a line of text in a paragraph.

Shaping

The shaping strategy of some text.

Wrapping

The wrapping strategy of some text.

Traits

IntoFragment

A trait for converting a value to some text Fragment.

Renderer

A renderer capable of measuring and drawing Text.

Type Aliases

Fragment

A fragment of Text.