iced_core::text::editor

Trait Editor

pub trait Editor: Sized + Default {
    type Font: Copy + PartialEq + Default;

    // Required methods
    fn with_text(text: &str) -> Self;
    fn is_empty(&self) -> bool;
    fn cursor(&self) -> Cursor;
    fn selection(&self) -> Selection;
    fn copy(&self) -> Option<String>;
    fn line(&self, index: usize) -> Option<Line<'_>>;
    fn line_count(&self) -> usize;
    fn perform(&mut self, action: Action);
    fn move_to(&mut self, cursor: Cursor);
    fn bounds(&self) -> Size;
    fn min_bounds(&self) -> Size;
    fn hint_factor(&self) -> Option<f32>;
    fn update(
        &mut self,
        new_bounds: Size,
        new_font: Self::Font,
        new_size: Pixels,
        new_line_height: LineHeight,
        new_wrapping: Wrapping,
        new_alignment: Alignment,
        new_hint_factor: Option<f32>,
        new_parser: &mut impl Parser,
    );
    fn overwrite(&mut self, new_text: &str);
    fn highlight<P: Parser>(
        &mut self,
        font: Self::Font,
        parser: &mut P,
        highlight: impl Fn(P::Output) -> Style,
    );
    fn font(&self) -> Self::Font;
    fn text_size(&self) -> Pixels;
    fn line_height(&self) -> LineHeight;

    // Provided methods
    fn lines(&self) -> impl Iterator<Item = Line<'_>> { ... }
    fn text(&self) -> String { ... }

}

A component that can be used by widgets to edit multi-line text.

Required Associated Types

type Font: Copy + PartialEq + Default

The font of the Editor.

Required Methods

fn with_text(text: &str) -> Self

Creates a new Editor laid out with the given text.

fn is_empty(&self) -> bool

Returns true if the Editor has no contents.

fn cursor(&self) -> Cursor

Returns the current Cursor of the Editor.

fn selection(&self) -> Selection

Returns the current Selection of the Editor.

fn copy(&self) -> Option<String>

Returns the current selected text of the Editor.

fn line(&self, index: usize) -> Option<Line<'_>>

Returns the text of the given line in the Editor, if it exists.

fn line_count(&self) -> usize

Returns the amount of lines in the Editor.

fn perform(&mut self, action: Action)

Performs an Action on the Editor.

fn move_to(&mut self, cursor: Cursor)

Moves the cursor to the given position.

fn bounds(&self) -> Size

Returns the current boundaries of the Editor.

fn min_bounds(&self) -> Size

Returns the minimum boundaries to fit the current contents of the Editor.

fn hint_factor(&self) -> Option<f32>

Returns the hint factor of the Editor.

fn update( &mut self, new_bounds: Size, new_font: Self::Font, new_size: Pixels, new_line_height: LineHeight, new_wrapping: Wrapping, new_alignment: Alignment, new_hint_factor: Option<f32>, new_parser: &mut impl Parser, )

Updates the Editor with some new attributes.

fn overwrite(&mut self, new_text: &str)

Overwrites the current contents of the Editor.

fn highlight<P: Parser>( &mut self, font: Self::Font, parser: &mut P, highlight: impl Fn(P::Output) -> Style, )

Runs a text::Highlighter in the Editor.

fn font(&self) -> Self::Font

Returns the current Font of the Editor.

fn text_size(&self) -> Pixels

Returns the current text size of the Editor.

fn line_height(&self) -> LineHeight

Returns the current LineHeight of the Editor.

Provided Methods

fn lines(&self) -> impl Iterator<Item = Line<'_>>

Returns an iterator of the text of the lines in the Editor.

fn text(&self) -> String

Returns the text of the Editor.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.