Trait iced_core::text::Paragraph

pub trait Paragraph: Default {
    type Font;

    // Required methods
    fn content(&self) -> &str;
    fn text_size(&self) -> Pixels;
    fn line_height(&self) -> LineHeight;
    fn font(&self) -> Self::Font;
    fn shaping(&self) -> Shaping;
    fn horizontal_alignment(&self) -> Horizontal;
    fn vertical_alignment(&self) -> Vertical;
    fn bounds(&self) -> Size;
    fn min_bounds(&self) -> Size;
    fn hit_test(&self, point: Point) -> Option<Hit>;
    fn grapheme_position(&self, line: usize, index: usize) -> Option<Point>;

    // Provided methods
    fn min_width(&self) -> f32 { ... }
    fn min_height(&self) -> f32 { ... }

}

A text paragraph.

Required Associated Types

type Font

The font of this Paragraph.

Required Methods

fn content(&self) -> &str

Returns the content of the Paragraph.

fn text_size(&self) -> Pixels

Returns the text size of the Paragraph.

fn line_height(&self) -> LineHeight

Returns the LineHeight of the Paragraph.

fn font(&self) -> Self::Font

Returns the Self::Font of the Paragraph.

fn shaping(&self) -> Shaping

Returns the Shaping strategy of the Paragraph.

fn horizontal_alignment(&self) -> Horizontal

Returns the horizontal alignment of the Paragraph.

fn vertical_alignment(&self) -> Vertical

Returns the vertical alignment of the Paragraph.

fn bounds(&self) -> Size

Returns the boundaries of the Paragraph.

fn min_bounds(&self) -> Size

Returns the minimum boundaries that can fit the contents of the Paragraph.

fn hit_test(&self, point: Point) -> Option<Hit>

Tests whether the provided point is within the boundaries of the Paragraph, returning information about the nearest character.

fn grapheme_position(&self, line: usize, index: usize) -> Option<Point>

Returns the distance to the given grapheme index in the Paragraph.

Provided Methods

fn min_width(&self) -> f32

Returns the minimum width that can fit the contents of the Paragraph.

fn min_height(&self) -> f32

Returns the minimum height that can fit the contents of the Paragraph.