iced_core::text

Trait Renderer

pub trait Renderer: Renderer {
    type Font: Copy + PartialEq;
    type Paragraph: Paragraph<Font = Self::Font> + 'static;
    type Editor: Editor<Font = Self::Font> + 'static;

    const ICON_FONT: Self::Font;
    const CHECKMARK_ICON: char;
    const ARROW_DOWN_ICON: char;
    const SCROLL_UP_ICON: char;
    const SCROLL_DOWN_ICON: char;
    const SCROLL_LEFT_ICON: char;
    const SCROLL_RIGHT_ICON: char;
    const ICED_LOGO: char;

    // Required methods
    fn default_font(&self) -> Self::Font;
    fn default_size(&self) -> Pixels;
    fn fill_paragraph(
        &mut self,
        text: &Self::Paragraph,
        position: Point,
        color: Color,
        clip_bounds: Rectangle,
    );
    fn fill_editor(
        &mut self,
        editor: &Self::Editor,
        position: Point,
        color: Color,
        clip_bounds: Rectangle,
    );
    fn fill_text(
        &mut self,
        text: Text<String, Self::Font>,
        position: Point,
        color: Color,
        clip_bounds: Rectangle,
    );
}

A renderer capable of measuring and drawing Text.

Required Associated Constants

const ICON_FONT: Self::Font

The icon font of the backend.

const CHECKMARK_ICON: char

The char representing a ✔ icon in the ICON_FONT.

const ARROW_DOWN_ICON: char

The char representing a ▼ icon in the built-in ICON_FONT.

const SCROLL_UP_ICON: char

The char representing a ^ icon in the built-in ICON_FONT.

const SCROLL_DOWN_ICON: char

The char representing a v icon in the built-in ICON_FONT.

const SCROLL_LEFT_ICON: char

The char representing a < icon in the built-in ICON_FONT.

const SCROLL_RIGHT_ICON: char

The char representing a > icon in the built-in ICON_FONT.

const ICED_LOGO: char

The ‘char’ representing the iced logo in the built-in ‘ICON_FONT’.

Required Associated Types

type Font: Copy + PartialEq

The font type used.

type Paragraph: Paragraph<Font = Self::Font> + 'static

The Paragraph of this Renderer.

type Editor: Editor<Font = Self::Font> + 'static

The Editor of this Renderer.

Required Methods

fn default_font(&self) -> Self::Font

Returns the default Self::Font.

fn default_size(&self) -> Pixels

Returns the default size of Text.

fn fill_paragraph( &mut self, text: &Self::Paragraph, position: Point, color: Color, clip_bounds: Rectangle, )

Draws the given Paragraph at the given position and with the given Color.

fn fill_editor( &mut self, editor: &Self::Editor, position: Point, color: Color, clip_bounds: Rectangle, )

Draws the given Editor at the given position and with the given Color.

fn fill_text( &mut self, text: Text<String, Self::Font>, position: Point, color: Color, clip_bounds: Rectangle, )

Draws the given Text at the given position and with the given Color.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.