iced_core::widget

Module operation

Query or update internal widget state.

Re-exports

pub use focusable::Focusable;

pub use scrollable::Scrollable;

pub use text_input::TextInput;

Modules

focusable

Operate on widgets that can be focused.

scrollable

Operate on widgets that can be scrolled.

search_id

Search for widgets with the target Id.

text_input

Operate on widgets that have text input.

Enums

Outcome

The result of an Operation.

Traits

Operation

A piece of logic that can traverse the widget tree of an application in order to query or update some widget state.

Functions

black_box

Wraps the Operation in a black box, erasing its returning type.

map

Maps the output of an Operation using the given function.

scoped

Produces an Operation that applies the given Operation to the children of a container with the given Id.

then

Chains the output of an Operation with the provided function to build a new Operation.