iced_core::widget::operation
Module focusable

Operate on widgets that can be focused.

Structs

Count

A summary of the focusable widgets present on a widget tree.

Traits

Focusable

The internal state of a widget that can be focused.

Functions

count

Produces an Operation that generates a Count and chains it with the provided function to build a new Operation.

find_focused

Produces an Operation that searches for the current focused widget and stores its ID. This ignores widgets that do not have an ID.

focus

Produces an Operation that focuses the widget with the given Id.

focus_next

Produces an Operation that searches for the current focused widget, and

focus_previous

Produces an Operation that searches for the current focused widget, and