iced_core::widget::operation::text_input

Trait TextInput

pub trait TextInput {
    // Required methods
    fn move_cursor_to_front(&mut self);
    fn move_cursor_to_end(&mut self);
    fn move_cursor_to(&mut self, position: usize);
    fn select_all(&mut self);
}

The internal state of a widget that has text input.

Required Methods

fn move_cursor_to_front(&mut self)

Moves the cursor of the text input to the front of the input text.

fn move_cursor_to_end(&mut self)

Moves the cursor of the text input to the end of the input text.

fn move_cursor_to(&mut self, position: usize)

Moves the cursor of the text input to an arbitrary location.

fn select_all(&mut self)

Selects all the content of the text input.