iced_core::widget

Module text

Text widgets display information through writing.

Example

use iced::widget::text;
use iced::color;

enum Message {
    // ...
}

fn view(state: &State) -> Element<'_, Message> {
    text("Hello, this is iced!")
        .size(20)
        .color(color!(0x0000ff))
        .into()
}

Re-exports

pub use text::LineHeight;

pub use text::Shaping;

pub use text::Wrapping;

Structs

State

The internal state of a Text widget.

Style

The appearance of some text.

Text

A bunch of text.

Traits

Catalog

The theme catalog of a Text.

Functions

base

Text with the default base color.

danger

Text conveying some negative information, like an error.

default

The default text styling; color is inherited.

draw

Draws text using the same logic as the Text widget.

layout

Produces the layout::Node of a Text widget.

primary

Text conveying some important information, like an action.

secondary

Text conveying some secondary information, like a footnote.

success

Text conveying some positive information, like a successful event.

Type Aliases

StyleFn

A styling function for a Text.