iced_core::widget::text

Function draw

pub fn draw<Renderer>(
    renderer: &mut Renderer,
    style: &Style,
    layout: Layout<'_>,
    paragraph: &Renderer::Paragraph,
    appearance: Style,
    viewport: &Rectangle,
)
where
    Renderer: Renderer,

Draws text using the same logic as the Text widget.

Specifically:

If no size is provided, the default text size of the Renderer will be used.

If no color is provided, the renderer::Style::text_color will be used.

The alignment attributes do not affect the position of the bounds of the Layout.