Function iced_core::widget::tree::diff_children_custom_with_search

pub fn diff_children_custom_with_search<T>(
    current_children: &mut Vec<Tree>,
    new_children: &[T],
    diff: impl Fn(&mut Tree, &T),
    maybe_changed: impl Fn(usize) -> bool,
    new_state: impl Fn(&T) -> Tree,
)

Reconciles the current_children with the provided list of widgets using custom logic both for diffing and creating new widget state.

The algorithm will try to minimize the impact of diffing by querying the maybe_changed closure.