iced_core
Module window

Build window-based GUI applications.

Re-exports

pub use icon::Icon;

pub use settings::Settings;

Modules

icon

Change the icon of a window.

settings

Configure your windows.

Structs

Id

The id of the window.

Enums

Event

A window-related event.

Level

A window level groups windows with respect to their z-position.

Mode

The mode of a window-based application.

Position

The position of a window in a given screen.

RedrawRequest

A request to redraw a window.

UserAttention

The type of user attention to request.