iced_core::window::settings

Struct Settings

pub struct Settings {

    pub size: Size,
    pub maximized: bool,
    pub fullscreen: bool,
    pub position: Position,
    pub min_size: Option<Size>,
    pub max_size: Option<Size>,
    pub visible: bool,
    pub resizable: bool,
    pub decorations: bool,
    pub transparent: bool,
    pub level: Level,
    pub icon: Option<Icon>,
    pub platform_specific: PlatformSpecific,
    pub exit_on_close_request: bool,

}

The window settings of an application.

Fields
size: Size

The initial logical dimensions of the window.

maximized: bool

Whether the window should start maximized.

fullscreen: bool

Whether the window should start fullscreen.

position: Position

The initial position of the window.

min_size: Option<Size>

The minimum size of the window.

max_size: Option<Size>

The maximum size of the window.

visible: bool

Whether the window should be visible or not.

resizable: bool

Whether the window should be resizable or not.

decorations: bool

Whether the window should have a border, a title bar, etc. or not.

transparent: bool

Whether the window should be transparent.

level: Level

The window Level.

icon: Option<Icon>

The icon of the window.

platform_specific: PlatformSpecific

Platform specific settings.

exit_on_close_request: bool

Whether the window will close when the user requests it, e.g. when a user presses the close button.

This can be useful if you want to have some behavior that executes before the window is actually destroyed. If you disable this, you must manually close the window with the window::close command.

By default this is enabled.