Struct iced_core::Shell

pub struct Shell<'a, Message> { /* private fields */ }

A connection to the state of a shell.

A Widget can leverage a Shell to trigger changes in an application, like publishing messages or invalidating the current layout.

Implementations

impl<'a, Message> Shell<'a, Message>

pub fn new(messages: &'a mut Vec<Message>) -> Self

Creates a new Shell with the provided buffer of messages.

pub fn is_empty(&self) -> bool

Returns true if the Shell contains no published messages

pub fn publish(&mut self, message: Message)

Publish the given Message for an application to process it.

pub fn request_redraw(&mut self, request: RedrawRequest)

Requests a new frame to be drawn.

pub fn redraw_request(&self) -> Option<RedrawRequest>

Returns the request a redraw should happen, if any.

pub fn is_layout_invalid(&self) -> bool

Returns whether the current layout is invalid or not.

pub fn invalidate_layout(&mut self)

Invalidates the current application layout.

The shell will relayout the application widgets.

pub fn revalidate_layout(&mut self, f: impl FnOnce())

Triggers the given function if the layout is invalid, cleaning it in the process.

pub fn are_widgets_invalid(&self) -> bool

Returns whether the widgets of the current application have been invalidated.

pub fn invalidate_widgets(&mut self)

Invalidates the current application widgets.

The shell will rebuild and relayout the widget tree.

pub fn merge<B>(&mut self, other: Shell<'_, B>, f: impl Fn(B) -> Message)

Merges the current Shell with another one by applying the given function to the messages of the latter.

This method is useful for composition.