Trait iced_core::text::Renderer

pub trait Renderer: Renderer {
    type Font: Copy;

    const ICON_FONT: Self::Font;
    const CHECKMARK_ICON: char;
    const ARROW_DOWN_ICON: char;

    // Required methods
    fn default_font(&self) -> Self::Font;
    fn default_size(&self) -> f32;
    fn measure(
        &self,
        content: &str,
        size: f32,
        line_height: LineHeight,
        font: Self::Font,
        bounds: Size,
        shaping: Shaping
    ) -> (f32, f32);
    fn hit_test(
        &self,
        contents: &str,
        size: f32,
        line_height: LineHeight,
        font: Self::Font,
        bounds: Size,
        shaping: Shaping,
        point: Point,
        nearest_only: bool
    ) -> Option<Hit>;
    fn load_font(&mut self, font: Cow<'static, [u8]>);
    fn fill_text(&mut self, text: Text<'_, Self::Font>);

    // Provided method
    fn measure_width(
        &self,
        content: &str,
        size: f32,
        font: Self::Font,
        shaping: Shaping
    ) -> f32 { ... }
}

A renderer capable of measuring and drawing Text.

Required Associated Types

type Font: Copy

The font type used.

Required Associated Constants

const ICON_FONT: Self::Font

The icon font of the backend.

const CHECKMARK_ICON: char

The char representing a ✔ icon in the ICON_FONT.

const ARROW_DOWN_ICON: char

The char representing a ▼ icon in the built-in ICON_FONT.

Required Methods

fn default_font(&self) -> Self::Font

Returns the default Self::Font.

fn default_size(&self) -> f32

Returns the default size of Text.

fn measure( &self, content: &str, size: f32, line_height: LineHeight, font: Self::Font, bounds: Size, shaping: Shaping ) -> (f32, f32)

Measures the text in the given bounds and returns the minimum boundaries that can fit the contents.

fn hit_test( &self, contents: &str, size: f32, line_height: LineHeight, font: Self::Font, bounds: Size, shaping: Shaping, point: Point, nearest_only: bool ) -> Option<Hit>

Tests whether the provided point is within the boundaries of text laid out with the given parameters, returning information about the nearest character.

If nearest_only is true, the hit test does not consider whether the the point is interior to any glyph bounds, returning only the character with the nearest centeroid.

fn load_font(&mut self, font: Cow<'static, [u8]>)

Loads a Self::Font from its bytes.

fn fill_text(&mut self, text: Text<'_, Self::Font>)

Draws the given Text.

Provided Methods

fn measure_width( &self, content: &str, size: f32, font: Self::Font, shaping: Shaping ) -> f32

Measures the width of the text as if it were laid out in a single line.