Trait iced_core::text::Renderer

pub trait Renderer: Renderer {
    type Font: Copy + PartialEq;
    type Paragraph: Paragraph<Font = Self::Font> + 'static;

    const ICON_FONT: Self::Font;
    const CHECKMARK_ICON: char;
    const ARROW_DOWN_ICON: char;

    // Required methods
    fn default_font(&self) -> Self::Font;
    fn default_size(&self) -> Pixels;
    fn load_font(&mut self, font: Cow<'static, [u8]>);
    fn create_paragraph(&self, text: Text<'_, Self::Font>) -> Self::Paragraph;
    fn resize_paragraph(
        &self,
        paragraph: &mut Self::Paragraph,
        new_bounds: Size
    );
    fn fill_paragraph(
        &mut self,
        text: &Self::Paragraph,
        position: Point,
        color: Color
    );
    fn fill_text(
        &mut self,
        text: Text<'_, Self::Font>,
        position: Point,
        color: Color
    );

    // Provided method
    fn update_paragraph(
        &self,
        paragraph: &mut Self::Paragraph,
        text: Text<'_, Self::Font>
    ) { ... }
}

A renderer capable of measuring and drawing Text.

Required Associated Types

type Font: Copy + PartialEq

The font type used.

type Paragraph: Paragraph<Font = Self::Font> + 'static

The Paragraph of this Renderer.

Required Associated Constants

const ICON_FONT: Self::Font

The icon font of the backend.

const CHECKMARK_ICON: char

The char representing a ✔ icon in the ICON_FONT.

const ARROW_DOWN_ICON: char

The char representing a ▼ icon in the built-in ICON_FONT.

Required Methods

fn default_font(&self) -> Self::Font

Returns the default Self::Font.

fn default_size(&self) -> Pixels

Returns the default size of Text.

fn load_font(&mut self, font: Cow<'static, [u8]>)

Loads a Self::Font from its bytes.

fn create_paragraph(&self, text: Text<'_, Self::Font>) -> Self::Paragraph

Creates a new Paragraph laid out with the given Text.

fn resize_paragraph(&self, paragraph: &mut Self::Paragraph, new_bounds: Size)

Lays out the given Paragraph with some new boundaries.

fn fill_paragraph( &mut self, text: &Self::Paragraph, position: Point, color: Color )

Draws the given Paragraph at the given position and with the given Color.

fn fill_text( &mut self, text: Text<'_, Self::Font>, position: Point, color: Color )

Draws the given Text at the given position and with the given Color.

Provided Methods

fn update_paragraph( &self, paragraph: &mut Self::Paragraph, text: Text<'_, Self::Font> )

Updates a Paragraph to match the given Text, if needed.