iced_core::widget::tree

Struct Tree

pub struct Tree {
    pub tag: Tag,
    pub state: State,
    pub children: Vec<Tree>,
}

A persistent state widget tree.

A Tree is normally associated with a specific widget in the widget tree.

Fields
tag: Tag

The tag of the Tree.

state: State

The State of the Tree.

children: Vec<Tree>

The children of the root widget of the Tree.

Implementations

impl Tree

pub fn empty() -> Self

Creates an empty, stateless Tree with no children.

pub fn new<'a, Message, Theme, Renderer>( widget: impl Borrow<dyn Widget<Message, Theme, Renderer> + 'a>, ) -> Self
where Renderer: Renderer,

Creates a new Tree for the provided Widget.

pub fn diff<'a, Message, Theme, Renderer>( &mut self, new: impl Borrow<dyn Widget<Message, Theme, Renderer> + 'a>, )
where Renderer: Renderer,

Reconciles the current tree with the provided Widget.

If the tag of the Widget matches the tag of the Tree, then the Widget proceeds with the reconciliation (i.e. Widget::diff is called).

Otherwise, the whole Tree is recreated.

pub fn diff_children<'a, Message, Theme, Renderer>( &mut self, new_children: &[impl Borrow<dyn Widget<Message, Theme, Renderer> + 'a>], )
where Renderer: Renderer,

Reconciles the children of the tree with the provided list of widgets.

pub fn diff_children_custom<T>( &mut self, new_children: &[T], diff: impl Fn(&mut Tree, &T), new_state: impl Fn(&T) -> Self, )

Reconciles the children of the tree with the provided list of widgets using custom logic both for diffing and creating new widget state.